# beta-private/api/NHSD.GPITF.BuyingCatalog/crm/Gif.Service/wwwroot/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "version" : "1.0.0-private-beta",
    "title" : "catalogue-api",
    "description" : "NHS Digital GP IT Futures Buying Catalog API"
  },
  "paths" : {
...
    "/api/Solutions/ByFramework/{frameworkId}" : {
      "get" : {
        "tags" : [ "Solutions" ],
        "summary" : "Get existing solution/s on which were onboarded onto a framework,\r\ngiven the CRM identifier of the framework",
        "operationId" : "ApiSolutionsByFrameworkByFrameworkIdGet",
        "consumes" : [ ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "frameworkId",
          "in" : "path",
          "description" : "CRM identifier of organisation to find",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "pageIndex",
          "in" : "query",
          "description" : "1-based index of page to return.  Defaults to 1",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "pageSize",
          "in" : "query",
          "description" : "number of items per page.  Defaults to 20",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/PaginatedList[Solutions]"
            }
          },
          "404" : {
            "description" : "Framework not found in CRM"
          }
        },
        "security" : [ {
          "Bearer" : [ ]
        } ]
      }
    },
    "/api/Solutions/ById/{id}" : {
      "get" : {
        "tags" : [ "Solutions" ],
        "summary" : "Get an existing solution given its CRM identifier\r\nTypically used to retrieve previous version",
        "operationId" : "ApiSolutionsByIdByIdGet",
        "consumes" : [ ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "CRM identifier of solution to find",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/Solutions"
            }
          },
          "404" : {
            "description" : "Solution not found in CRM"
          }
        },
        "security" : [ {
          "Bearer" : [ ]
        } ]
      }
    },
    "/api/Solutions/ByOrganisation/{organisationId}" : {
      "get" : {
        "tags" : [ "Solutions" ],
        "summary" : "Retrieve all current solutions in a paged list for an organisation,\r\ngiven the organisation’s CRM identifier",
        "operationId" : "ApiSolutionsByOrganisationByOrganisationIdGet",
        "consumes" : [ ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "organisationId",
          "in" : "path",
          "description" : "CRM identifier of organisation",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "pageIndex",
          "in" : "query",
          "description" : "1-based index of page to return.  Defaults to 1",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "pageSize",
          "in" : "query",
          "description" : "number of items per page.  Defaults to 20",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/PaginatedList[Solutions]"
            }
          },
          "404" : {
            "description" : "Organisation not found in CRM"
          }
        },
        "security" : [ {
          "Bearer" : [ ]
        } ]
      }
    },
    "/api/Solutions" : {
      "put" : {
        "tags" : [ "Solutions" ],
        "summary" : "Update an existing solution with new information",
        "operationId" : "ApiSolutionsPut",
        "consumes" : [ "application/json-patch+json", "application/json", "text/json", "application/*+json" ],
        "produces" : [ ],
        "parameters" : [ {
          "name" : "solution",
          "in" : "body",
          "description" : "contact with updated information",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Solutions"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success"
          },
          "404" : {
            "description" : "Organisation or solution not found in CRM"
          }
        },
        "security" : [ {
          "Bearer" : [ ]
        } ]
      },
      "post" : {
        "tags" : [ "Solutions" ],
        "summary" : "Create a new solution for an organisation",
        "operationId" : "ApiSolutionsPost",
        "consumes" : [ "application/json-patch+json", "application/json", "text/json", "application/*+json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "solution",
          "in" : "body",
          "description" : "new solution information",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Solutions"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/Solutions"
            }
          },
          "404" : {
            "description" : "Organisation not found in CRM"
          },
          "500" : {
            "description" : "Validation exception"
          }
        },
        "security" : [ {
          "Bearer" : [ ]
        } ]
      },
    "delete" : {
        "tags" : [ "Solutions" ],
        "summary" : "Delete an existing solution (will cascade delete to child entities)",
        "operationId" : "ApiSolutionsDelete",
        "consumes" : [ "application/json-patch+json", "application/json", "text/json", "application/*+json" ],
        "produces" : [ ],
        "parameters" : [ {
          "name" : "solution",
          "in" : "body",
          "description" : "existing solution id",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Solutions"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success"
          },
          "404" : {
            "description" : "Solution not found in CRM"
          }
        },
        "security" : [ {
          "Bearer" : [ ]
        } ]
      }
    },
...
    "/api/StandardsApplicable" : {
      "put" : {
        "tags" : [ "StandardsApplicable" ],
        "summary" : "Update an existing claimed standard with new information",
        "operationId" : "ApiStandardsApplicablePut",
        "consumes" : [ "application/json-patch+json", "application/json", "text/json", "application/*+json" ],
        "produces" : [ ],
        "parameters" : [ {
          "name" : "claimedstandard",
          "in" : "body",
          "description" : "claimed standard with updated information",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/StandardsApplicable"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success"
          },
          "404" : {
            "description" : "Solution or ClaimedStandard not found in CRM"
          }
        },
        "security" : [ {
          "Bearer" : [ ]
        } ]
      },
      "post" : {
        "tags" : [ "StandardsApplicable" ],
        "summary" : "Create a new claimed standard for a solution",
        "operationId" : "ApiStandardsApplicablePost",
        "consumes" : [ "application/json-patch+json", "application/json", "text/json", "application/*+json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "claimedstandard",
          "in" : "body",
          "description" : "new claimed standard information",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/StandardsApplicable"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/StandardsApplicable"
            }
          },
          "404" : {
            "description" : "Solution not found in CRM"
          }
        },
        "security" : [ {
          "Bearer" : [ ]
        } ]
      },
      "delete" : {
        "tags" : [ "StandardsApplicable" ],
        "summary" : "Delete an existing claimed standard for a solution",
        "operationId" : "ApiStandardsApplicableDelete",
        "consumes" : [ "application/json-patch+json", "application/json", "text/json", "application/*+json" ],
        "produces" : [ ],
        "parameters" : [ {
          "name" : "claimedstandard",
          "in" : "body",
          "description" : "existing claimed standard information",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/StandardsApplicable"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success"
          },
          "404" : {
            "description" : "ClaimedStandard not found in CRM"
          }
        },
        "security" : [ {
          "Bearer" : [ ]
        } ]
      }
    },
...
  },
  "definitions" : {
...
    "Solutions" : {
      "description" : "A product and/or service provided by an ‘organisation’.\r\nNote that a ‘solution’ has a link to zero or one previous ‘solution’\r\nGenerally, only interested in current ‘solution’\r\nStandard MS Dynamics CRM entity",
      "required" : [ "id", "organisationId", "createdById", "modifiedById" ],
      "type" : "object",
      "properties" : {
        "id" : {
          "description" : "Unique identifier of entity",
          "type" : "string"
        },
        "previousId" : {
          "description" : "Unique identifier of previous version of entity",
          "type" : "string"
        },
        "organisationId" : {
          "description" : "Unique identifier of organisation",
          "type" : "string"
        },
        "version" : {
          "description" : "Version of solution",
          "type" : "string"
        },
        "status" : {
          "description" : "Current status of this solution",
          "enum" : [ "Failed", "Draft", "Registered", "CapabilitiesAssessment", "StandardsCompliance", "FinalApproval", "SolutionPage", "Approved" ],
          "type" : "string"
        },
        "createdById" : {
          "description" : "Unique identifier of Contact who created this entity\r\nDerived from calling context\r\nSET ON SERVER",
          "type" : "string"
        },
        "createdOn" : {
          "format" : "date-time",
          "description" : "UTC date and time at which record created\r\nSet by server when creating record\r\nSET ON SERVER",
          "type" : "string"
        },
        "modifiedById" : {
          "description" : "Unique identifier of Contact who last modified this entity\r\nDerived from calling context\r\nSET ON SERVER",
          "type" : "string"
        },
        "modifiedOn" : {
          "format" : "date-time",
          "description" : "UTC date and time at which record last modified\r\nSet by server when creating record\r\nSET ON SERVER",
          "type" : "string"
        },
        "name" : {
          "description" : "Name of solution, as displayed to a user",
          "type" : "string"
        },
        "description" : {
          "description" : "Description of solution, as displayed to a user",
          "type" : "string"
        },
        "productPage" : {
          "description" : "Serialised product page",
          "type" : "string"
        }
      },
      "example" : {
        "id" : "A3C6830F-2E7C-4545-A4B9-02D20C4C92E1",
        "previousId" : "12968eb4-4160-4ec5-8bb7-3deca7c3f53b",
        "organisationId" : "9c2fd4d4-98ca-41d7-b8f4-1ad6f6702127",
        "version" : "2.0",
        "status" : "Submitted",
        "createdById" : "371b20b7-3bed-487d-aa5f-995f5924e579",
        "createdOn" : "2017-12-25T06:45:00",
        "modifiedById" : "371b20b7-3bed-487d-aa5f-995f5924e579",
        "modifiedOn" : "2018-12-31T17:25:00",
        "name" : "Really Kool Document Manager",
        "description" : "Does Really Kool document management",
        "productPage" : ""
      }
    },
    "StandardsApplicable" : {
      "description" : "A ‘Standard’ which a ‘Solution’ asserts that it provides.\r\nThis is then assessed by NHS to verify the ‘Solution’ complies with the ‘Standard’ it has claimed.",
      "required" : [ "standardId", "id", "solutionId" ],
      "type" : "object",
      "properties" : {
        "standardId" : {
          "description" : "Unique identifier of standard",
          "type" : "string"
        },
        "status" : {
          "description" : "Current status of this ClaimedStandard",
          "enum" : [ "NotStarted", "Draft", "Submitted", "Remediation", "Approved", "ApprovedFirstOfType", "ApprovedPartial", "Rejected" ],
          "type" : "string"
        },
        "id" : {
          "description" : "Unique identifier of entity",
          "type" : "string"
        },
        "solutionId" : {
          "description" : "Unique identifier of solution",
          "type" : "string"
        }
      },
      "example" : {
        "standardId" : "INT2",
        "status" : "Submitted",
        "id" : "0F2614F9-2521-414A-A448-0096C0DF3ABE",
        "solutionId" : "A3C6830F-2E7C-4545-A4B9-02D20C4C92E1"
      }
    },
...
    "PaginatedList[Solutions]" : {
      "description" : "A paged list of objects",
      "type" : "object",
      "properties" : {
        "pageIndex" : {
          "format" : "int32",
          "description" : "1-based index of which page this page\r\nDefaults to 1",
          "type" : "integer",
          "readOnly" : true
        },
        "totalPages" : {
          "format" : "int32",
          "description" : "Total number of pages based on NHSD.GPITF.BuyingCatalog.Models.PaginatedList`1.PageSize",
          "type" : "integer",
          "readOnly" : true
        },
        "pageSize" : {
          "format" : "int32",
          "description" : "Maximum number of items in this page\r\nDefaults to 20",
          "type" : "integer",
          "readOnly" : true
        },
        "items" : {
          "description" : "List of items",
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Solutions"
          },
          "readOnly" : true
        }
      }
    },
...
  },
  "securityDefinitions" : {
    "Bearer" : {
      "type" : "apiKey",
      "name" : "Authorization",
      "in" : "header"
    }
  }
}